requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-telegram-bot==21.7
firebase-admin==6.6.0
google-cloud-firestore==2.19.0
google-genai==1.0.0
anthropic>=0.40.0
openai>=1.50.0
aiohttp>=3.9.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pydantic==2.10.0
python-dotenv==1.0.1
mem0ai>=0.1.30
